strndup

+++ includes/libft.h
@@ -45,7 +45,7 @@ int					ft_strcchr(char const *s, char c);
 int					ft_strclchr(char const *s, char c);
 int		ft_strrmchar(char *str, int c);
 char				*ft_strrmstr(char *str, char *torm);
-char				*ft_strndup(char const *str, int len);
+char    *ft_strndup(char const *str, size_t len);
 char				*ft_strinv(char *s);
 int					ft_strinsert(char *target, int pos, char *new);
 void				*ft_realloc(void **src, size_t src_size, size_t final_size);
